requirements.txt
datasets==4.0.0
ipykernel==6.30.1
langchain==0.3.27
pip==22.0.4
chromadb
python-dotenv
scikit-learn
pandas
numpy==1.26.4


transformers
sentence-transformers
llama-index==0.13.6
llama-index-embeddings-huggingface==0.6.0
langchain-openai
langchain-community
openai
nltk